time_tracker_project/requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.14.0
pip==24.0
pytz==2024.1
setuptools==65.5.0
sqlparse==0.4.4
tzdata==2024.1
django-cors-headers